# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Simple Pizza
host: localhost:10010
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  APIKeySecurity:
    type: apiKey
    in: header
    name: X-Session-ID
security:
  - APIKeySecurity: []
paths:
  /signup:
    x-swagger-router-controller: auth
    post:
      operationId: 'createUser'
      summary: Registers a User
      security: []
      parameters:
        - in: 'body'
          name: 'signup'
          required: true
          schema:
            $ref: '#/definitions/Credentials'
      responses:
        "200":  
          description: User registered
          schema:
            $ref: "#/definitions/UserResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/DefaultResponse"
...
definitions:
  Credentials:
    required:
      - email
      - password
    properties:
      email:
        type: string
      password:
        type: string
      name:
        type: string
  DefaultResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  UserResponse:
    required:
      - id
      - email
      - password
      - money
      - coupon
    properties:
      id:
        type: string
      email:
        type: string
      password:
        type: string
      money:
        type: integer
      coupon:
        type: integer